[F] Add QueryState context and hook
Currently, we have a global "loading" state in PageContext. When the
QueryWrapper is loading, it sets this loading state, which triggers the
appearance of the global ProgressBar component.

With the goal of improving UX, we're trying to let smaller components
make decisions about how to render before their data has arrived instead
of blocking rendering at a high level until data is present. To
accomplish this, each component wrapped by QueryWrapper may need to know
the current loading state. This state can be started, loading, or
completed, and for now we're capturing it in separate boolean variables
rather than a state machine to keep the code more expressive. This could
change as our needs become more complex.

Consider the exampe of a ModelList when thinking about why we need three
states. The ModelList needs to display a loading state before the query
has started and while it's loading. Prior to introducing these states,
the ModelList would initially show a "no results" message, then replace
it with a loading state because it was behaving as though the query was
cmoplete prior to the query being fired.