[F] Improve ModelTable and ModelList loading UX
This commit adjusts ModelTable to determine its loading state from
QueryStateContext and render a loading state if the query is not yet
completed.

It adjusts ModelList so that it can distinguish between data not yet
being present and data being present and not having any rows. Prior to
this change, ModelList would flash a no results message before rendering
a loading indicator.