[C] Update typescript eslint
Additionally, pin typescript-estree to address warning when linting.
Currently, eslint-config-next is pulling in an older version of
@typescript-eslint/typescript-estree, and upgrading eslint-config-next
is difficult because it conflicts with the shared CIC ESLint config. We
may be able to remove this resolution and dev dependency once the CIC
ESLint module moves its dependencies to peer dependencies.